refactor: split orchestrator-routes.ts into focused sub-routers
Split the 706-line monolithic createOrchestratorRouter() into three
focused sub-routers, each handling a distinct concern:

- orchestrator-session-router.ts: session lifecycle, child sessions,
  reports, dashboard, pending prompts, and session cleanup
- orchestrator-memory-router.ts: memory CRUD, trust records, and
  notifications
- orchestrator-learning-router.ts: finding outcomes, skill levels,
  decisions, and memory extraction/aging

The original orchestrator-routes.ts is now a thin wrapper that mounts
all three sub-routers, preserving the existing export interface.
