+++ server/orchestrator-session-router.ts
@@ -1,0 +1,326 @@
+/**
+ * Session lifecycle, child session management, reports, and dashboard routes
+ * for the orchestrator.
+ */
+
+import { Router } from 'express'
+import type { Request } from 'express'
+import { resolve } from 'path'
+import { existsSync, statSync } from 'fs'
+import type { SessionManager } from './session-manager.js'
+import { ensureOrchestratorRunning, getOrchestratorSessionId, getOrCreateOrchestratorId } from './orchestrator-manager.js'
+import { getAgentDisplayName, REPOS_ROOT } from './config.js'
+import { scanRepoReports, readReport, getReportsSince } from './orchestrator-reports.js'
+import type { OrchestratorMemory } from './orchestrator-memory.js'
+import type { OrchestratorChildManager } from './orchestrator-children.js'
+import type { OrchestratorMonitor } from './orchestrator-monitor.js'
+
+// ---------------------------------------------------------------------------
+// Request body interfaces
+// ---------------------------------------------------------------------------
+
+interface SpawnChildBody {
+  repo: string
+  task: string
+  branchName: string
+  completionPolicy?: 'pr' | 'merge' | 'commit-only'
+  deployAfter?: boolean
+  useWorktree?: boolean
+  model?: string
+  allowedTools?: string[]
+}
+
+interface SessionRespondBody {
+  requestId?: string
+  value: string
+}
+
+// ---------------------------------------------------------------------------
+// Router factory
+// ---------------------------------------------------------------------------
+
+export function createSessionRouter(
+  verifyOrchestratorAuth: (req: Request) => boolean,
+  sessions: SessionManager,
+  memory: OrchestratorMemory,
+  children: OrchestratorChildManager,
+  monitorRef?: { current: OrchestratorMonitor | null },
+): Router {
+  const router = Router()
+
+  // -------------------------------------------------------------------------
+  // Session lifecycle
+  // -------------------------------------------------------------------------
+
+  /** Get orchestrator session status. */
+  router.get('/api/orchestrator/status', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const sessionId = getOrchestratorSessionId(sessions)
+    if (!sessionId) {
+      return res.json({ sessionId: null, status: 'stopped', agentName: getAgentDisplayName() })
+    }
+
+    const session = sessions.get(sessionId)
+    const status = session?.claudeProcess?.isAlive() ? 'active' : 'idle'
+    res.json({
+      sessionId,
+      status,
+      childSessions: children.activeCount(),
+      agentName: getAgentDisplayName(),
+    })
+  })
+
+  /** Ensure orchestrator is running and return its session ID. */
+  router.post('/api/orchestrator/start', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    try {
+      const sessionId = ensureOrchestratorRunning(sessions)
+      res.json({ sessionId, status: 'active', agentName: getAgentDisplayName() })
+    } catch (err) {
+      console.error('[orchestrator] Failed to start:', err)
+      res.status(500).json({ error: `Failed to start Agent ${getAgentDisplayName()}` })
+    }
+  })
+
+  // -------------------------------------------------------------------------
+  // Reports
+  // -------------------------------------------------------------------------
+
+  /** Scan reports for a single repo. */
+  router.get('/api/orchestrator/reports', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const repoPath = req.query.repo as string | undefined
+    const since = req.query.since as string | undefined
+
+    if (repoPath) {
+      const reports = scanRepoReports(repoPath)
+      res.json({ reports })
+    } else if (since) {
+      const repoItems = memory.list({ memoryType: 'repo_context' })
+      const repoPaths = repoItems.map(r => r.scope).filter((s): s is string => !!s)
+      const reports = getReportsSince(repoPaths, since)
+      res.json({ reports })
+    } else {
+      res.status(400).json({ error: 'Provide ?repo=<path> or ?since=<YYYY-MM-DD>' })
+    }
+  })
+
+  /** Read a specific report's content. */
+  router.get('/api/orchestrator/reports/read', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const filePath = req.query.path as string | undefined
+    if (!filePath) return res.status(400).json({ error: 'Provide ?path=<filePath>' })
+
+    const report = readReport(filePath)
+    if (!report) return res.status(404).json({ error: 'Report not found' })
+
+    res.json({ report })
+  })
+
+  // -------------------------------------------------------------------------
+  // Child sessions
+  // -------------------------------------------------------------------------
+
+  /** List child sessions. */
+  router.get('/api/orchestrator/children', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    res.json({ children: children.list() })
+  })
+
+  /** Spawn a child session. */
+  router.post('/api/orchestrator/children', async (req: Request<Record<string, string>, unknown, SpawnChildBody>, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const { repo, task, branchName, completionPolicy, deployAfter, useWorktree, model, allowedTools } = req.body
+    if (!repo || !task || !branchName) {
+      return res.status(400).json({ error: 'Missing required fields: repo, task, branchName' })
+    }
+
+    // Validate branchName to prevent prompt injection
+    if (!/^[a-zA-Z0-9][a-zA-Z0-9/_.-]*$/.test(branchName)) {
+      return res.status(400).json({ error: 'Invalid branchName: only alphanumeric, /, _, ., and - are allowed' })
+    }
+
+    // Validate allowedTools if provided: must be an array of strings
+    if (allowedTools !== undefined) {
+      if (!Array.isArray(allowedTools) || !allowedTools.every((t: unknown) => typeof t === 'string')) {
+        return res.status(400).json({ error: 'Invalid allowedTools: must be an array of strings' })
+      }
+    }
+
+    // Validate repo path: must resolve under REPOS_ROOT and be an existing directory
+    const resolvedRepo = resolve(repo)
+    if (!resolvedRepo.startsWith(REPOS_ROOT + '/') && resolvedRepo !== REPOS_ROOT) {
+      return res.status(400).json({ error: 'Invalid repo path: must be under configured repos root' })
+    }
+    if (!existsSync(resolvedRepo) || !statSync(resolvedRepo).isDirectory()) {
+      return res.status(400).json({ error: 'Invalid repo path: directory does not exist' })
+    }
+
+    try {
+      const child = await children.spawn({
+        repo,
+        task,
+        branchName,
+        completionPolicy: completionPolicy ?? 'pr',
+        deployAfter: deployAfter ?? false,
+        useWorktree: useWorktree ?? true,
+        model,
+        allowedTools,
+      })
+      res.json({ child })
+    } catch (err) {
+      res.status(503).json({ error: err instanceof Error ? err.message : 'Failed to spawn child session' })
+    }
+  })
+
+  /** Get a specific child session. */
+  router.get('/api/orchestrator/children/:id', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const child = children.get(req.params.id)
+    if (!child) return res.status(404).json({ error: 'Child session not found' })
+
+    res.json({ child })
+  })
+
+  // -------------------------------------------------------------------------
+  // Session prompts & approvals
+  // -------------------------------------------------------------------------
+
+  /** Get all sessions with pending prompts (waiting for approval or answer). */
+  router.get('/api/orchestrator/sessions/pending-prompts', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    res.json({ sessions: sessions.getPendingPrompts() })
+  })
+
+  /** Approve or deny a pending prompt in any session. */
+  router.post('/api/orchestrator/sessions/:id/respond', (req: Request<{ id: string }, unknown, SessionRespondBody>, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const sessionId = req.params.id
+    const { requestId, value } = req.body
+    if (!value) {
+      return res.status(400).json({ error: 'Missing required field: value (e.g. "allow", "deny", or answer text)' })
+    }
+
+    const session = sessions.get(sessionId)
+    if (!session) return res.status(404).json({ error: 'Session not found' })
+
+    // Verify there's actually a pending prompt (optionally for the specific requestId)
+    const hasPending = requestId
+      ? (session.pendingToolApprovals.has(requestId) || session.pendingControlRequests.has(requestId))
+      : (session.pendingToolApprovals.size > 0 || session.pendingControlRequests.size > 0)
+
+    if (!hasPending) {
+      return res.status(409).json({ error: 'No pending prompt to respond to' })
+    }
+
+    // Capture prompt details before responding (response clears them)
+    let promptToolName = 'unknown'
+    let promptType: 'permission' | 'question' = 'permission'
+    if (requestId) {
+      const toolApproval = session.pendingToolApprovals.get(requestId)
+      const controlReq = session.pendingControlRequests.get(requestId)
+      if (toolApproval) {
+        promptToolName = toolApproval.toolName
+        promptType = toolApproval.toolName === 'AskUserQuestion' ? 'question' : 'permission'
+      } else if (controlReq) {
+        promptToolName = controlReq.toolName
+        promptType = controlReq.toolName === 'AskUserQuestion' ? 'question' : 'permission'
+      }
+    }
+
+    sessions.sendPromptResponse(sessionId, value, requestId)
+
+    // Broadcast a notification to the orchestrator channel
+    const orchestratorId = getOrCreateOrchestratorId()
+    const orchestratorSession = sessions.get(orchestratorId)
+    if (orchestratorSession && orchestratorSession.clients.size > 0) {
+      const actionLabel = promptType === 'question'
+        ? `answered question from ${promptToolName}`
+        : `responded "${value}" to ${promptToolName}`
+      const notifMsg = {
+        type: 'system_message' as const,
+        subtype: 'info' as const,
+        text: `[${getAgentDisplayName()}] ${actionLabel} in session "${session.name}"`,
+      }
+      for (const ws of orchestratorSession.clients) {
+        ws.send(JSON.stringify(notifMsg))
+      }
+    }
+
+    res.json({ ok: true })
+  })
+
+  // -------------------------------------------------------------------------
+  // Session cleanup & listing
+  // -------------------------------------------------------------------------
+
+  /** List all sessions (unfiltered, includes source field). */
+  router.get('/api/orchestrator/sessions', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    res.json({ sessions: sessions.listAll() })
+  })
+
+  /** Delete all automated sessions (source: workflow, webhook, stepflow, agent). */
+  router.delete('/api/orchestrator/sessions/cleanup', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const automatedSources = new Set(['workflow', 'webhook', 'stepflow', 'agent'])
+    const toDelete = sessions.listAll().filter((s) => automatedSources.has(s.source ?? ''))
+
+    let deleted = 0
+    for (const s of toDelete) {
+      if (sessions.delete(s.id)) deleted++
+    }
+
+    res.json({ deleted })
+  })
+
+  /** Delete a specific session by ID. */
+  router.delete('/api/orchestrator/sessions/:id', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const success = sessions.delete(req.params.id)
+    if (!success) return res.status(404).json({ error: 'Session not found' })
+
+    res.json({ deleted: true })
+  })
+
+  // -------------------------------------------------------------------------
+  // Dashboard stats
+  // -------------------------------------------------------------------------
+
+  /** Get summary stats for the dashboard header. */
+  router.get('/api/orchestrator/dashboard', (req, res) => {
+    if (!verifyOrchestratorAuth(req)) return res.status(401).json({ error: 'Unauthorized' })
+
+    const repoItems = memory.list({ memoryType: 'repo_context' })
+    const pendingNotifications = monitorRef?.current?.getPending() ?? []
+    const activeChildren = children.activeCount()
+    const trustRecords = memory.listTrustRecords()
+    const autoApproved = trustRecords.filter(t => t.effectiveLevel !== 'ask').length
+
+    res.json({
+      stats: {
+        managedRepos: repoItems.length,
+        pendingNotifications: pendingNotifications.length,
+        activeChildSessions: activeChildren,
+        totalChildSessions: children.list().length,
+        trustRecords: trustRecords.length,
+        autoApprovedActions: autoApproved,
+        memoryItems: memory.list().length,
+      },
+    })
+  })
+
+  return router
+}
